round, don't truncate frame coordinate. keep consistent data type. clean comments
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@29489 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ display/d.save/main.c
@@ -72,7 +72,7 @@ int main (int argc, char **argv)
 	struct GModule *module;
 	char buff[1024];
 	char current_frame[64];
-	double Ftop, Fbot, Fleft, Fright;
+	float Ftop, Fbot, Fleft, Fright;
 
 
 	G_gisinit(argv[0]);
